feat(settings): consolidate MCP integrations into Connected accounts
Merge the MCPs settings page into Connected accounts so every external
connection lives in one place. The Integrations section lists the user's
MCP servers and surfaces featured not-yet-connected integrations (Linear,
Slack, Notion, PostHog) inline with one-click Connect, instead of hiding
them all in the Add server dialog. The old MCPs tab slims down to just the
Macro MCP server client setup instructions, relabeled "MCP server".

Also repoint the chat empty-state and sidebar quick links at the
consolidated page, and drop the now-unused agent settings sub-tab signal.

The featured-row connect flow chains add -> start auth via mutateAsync:
the cache update from the add unmounts the suggestion row, and
mutate()-level callbacks are dropped for unmounted observers.
