+++ js/app/packages/app/component/settings/ConnectedAccounts.tsx
@@ -3,10 +3,12 @@ import { ENABLE_EMAIL } from '@core/constant/featureFlags';
 import { SettingsPage, SettingsSection } from './primitives';
 import { EmailCard } from './Email';
 import { GitHubCard } from './GitHub';
+import { IntegrationsSection } from './Integrations';
 
 /**
- * Consolidated "Connected accounts" page: one card per external service the
- * user can link (Gmail, GitHub). Replaces the separate Email / GitHub tabs.
+ * Consolidated "Connected accounts" page: one card per external account the
+ * user can link (Gmail, GitHub), plus the agent's MCP integrations — so
+ * everything Macro is connected to lives in one place.
  */
 export function ConnectedAccounts() {
   return (
@@ -14,7 +16,7 @@ export function ConnectedAccounts() {
       title="Connected accounts"
       description="Connect your accounts so Macro can work across the tools you already use."
     >
-      <SettingsSection>
+      <SettingsSection title="Accounts">
         <div class="flex flex-col gap-3">
           <Show when={ENABLE_EMAIL}>
             <Suspense>
@@ -26,6 +28,9 @@ export function ConnectedAccounts() {
           </Suspense>
         </div>
       </SettingsSection>
+      <Suspense>
+        <IntegrationsSection />
+      </Suspense>
     </SettingsPage>
   );
 }
